WIP - testing DI
