+++ View/TestingSomething.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace oat\generis\View;
+
+/**
+ * Class TestingSomething
+ *
+ * @author Martijn Swinkels <m.swinkels@taotesting.com>
+ */
+class TestingSomething
+{
+    /**
+     * @return string
+     */
+    public function testing()
+    {
+        return 'Resolved!';
+    }
+}
