--- register.php
@@ -1,5 +1,0 @@
-<?php
-
-use oat\generis\Model\Extension\Registrar;
-
-Registrar::register('generis', __DIR__);
